Legacy -> frost-object-browser-inline, slots -> frost-object-browser
