+++ addon/styles/addon.scss
@@ -4,6 +4,43 @@
   display: flex;
   flex-direction: column;
 
+  &-body {
+    flex: 1;
+    display: flex;
+    flex-direction: row;
+    padding: 0 20px 0 30px;
+    background-color: $frost-color-lgrey-3;
+  }
+
+  &-facets {
+    margin: 20px 10px 10px 0;
+  }
+
+  &-content {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+  }
+
+  &-actions {
+    margin-right: 15px;
+    padding: 10px 16px 10px 0;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    align-items: center;
+
+    .frost-button,
+    .frost-link {
+      margin-left: 5px;
+    }
+  }
+}
+
+.frost-object-browser-inline {
+  display: flex;
+  flex-direction: column;
+
   .frost-list-item {
     .select-checkbox {
       .frost-checkbox {
@@ -183,42 +220,5 @@
 
   .no-display {
     display: none;
-  }
-}
-
-.frost-object-browser-slots {
-  display: flex;
-  flex-direction: column;
-
-  &-body {
-    flex: 1;
-    display: flex;
-    flex-direction: row;
-    padding: 0 20px 0 30px;
-    background-color: $frost-color-lgrey-3;
-  }
-
-  &-facets {
-    margin: 20px 10px 10px 0;
-  }
-
-  &-content {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-  }
-
-  &-actions {
-    margin-right: 15px;
-    padding: 10px 16px 10px 0;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: center;
-
-    .frost-button,
-    .frost-link {
-      margin-left: 5px;
-    }
   }
 }
