+++ addon/templates/components/frost-object-browser-inline.hbs
@@ -1,0 +1,89 @@
+<div class="frost-info-bar">
+  <div class="title">
+    <div class="primary-title">
+      {{title}}
+    </div>
+    <div class="sub-title">
+      {{computedSubtitle}}
+    </div>
+  </div>
+  <div class="top-buttons" background-color='red'>
+    {{frost-button
+      icon="frost/infobar-create"
+      onClick=(action "onCreate")
+      priority="tertiary"
+      size="medium"
+      text="Create"
+      vertical=true
+    }}
+  </div>
+</div>
+<div class="body">
+    {{#if filters}}
+    {{frost-object-browser-filter filters=filters onFilter=onFilter}}
+    {{/if}}
+  <div class="content">
+    <div class="tools">
+      <div class="sort"/>
+      <div class="pagination">
+        {{frost-object-browser-paginator
+          itemsPerPage=itemsPerPage
+          onPageChanged=(action 'onPageChanged')
+          page=computedPageNumber
+          total=computedValuesTotal
+        }}
+      </div>
+      <div class="detail">
+        <span class="text">
+          <!-- TODO: make this a variable to support i18n better -->
+          View:
+        </span>
+        <div class="button-bar {{ detailLevel }}">
+          {{#if viewSchema.low}}
+            {{frost-button
+              disabled=(eq detailLevel 'low')
+              onClick=(action 'onDetailChange' 'low')
+              priority="tertiary"
+              size="small"
+              icon="frost/list-small"
+            }}
+          {{/if}}
+          {{#if viewSchema.medium}}
+            {{frost-button
+              disabled=(eq detailLevel 'medium')
+              onClick=(action 'onDetailChange' 'medium')
+              priority="tertiary"
+              size="small"
+              icon="frost/list-medium"
+            }}
+          {{/if}}
+          {{#if viewSchema.high}}
+            {{frost-button
+              disabled=(eq detailLevel 'high')
+              onClick=(action 'onDetailChange' 'high')
+              priority="tertiary"
+              size="small"
+              icon="frost/list-large"
+            }}
+          {{/if}}
+        </div>
+      </div>
+    </div>
+    {{#frost-list onSelect=(action 'onSelect') selections=selectedItems records=computedValues as |record|}}
+      {{#frost-object-browser-list-item model=record as |value|}}
+        {{frost-bunsen-detail
+          bunsenModel=model
+          renderers=renderers
+          value=value
+          bunsenView=computedViewLevel
+        }}
+      {{/frost-object-browser-list-item}}
+    {{/frost-list}}
+    {{yield}}
+    <div class="actions">
+      {{#yield-slot 'actions'}}
+        {{yield 'actions'}}
+      {{/yield-slot}}
+    </div>
+  </div>
+</div>
