--- addon/templates/components/frost-object-browser-slots.hbs
@@ -1,44 +1,0 @@
-{{yield}}
-
-{{#yield-slot 'info-bar'}}
-  {{yield}}
-{{/yield-slot}}
-
-<div class='frost-object-browser-slots-body'>
-  {{#yield-slot 'facets'}}
-    <div class='frost-object-browser-slots-facets'>
-      {{yield}}
-    </div>
-  {{/yield-slot}}
-
-  <div class='frost-object-browser-slots-content'>
-    {{#yield-slot 'view'}}
-      {{yield
-        (block-params
-          (action 'onSelect')
-          selections
-        )
-      }}
-    {{/yield-slot}}
-
-    {{#yield-slot 'actions'}}
-      <div class='frost-object-browser-slots-actions'>
-        {{yield
-          (block-params
-            (hash
-              button=(component 'selection-action-button'
-                selections=selections
-                size='large'
-              )
-              link=(component 'selection-action-link'
-                selections=selections
-                size='large'
-              )
-            )
-            selections
-          )
-        }}
-      </div>
-    {{/yield-slot}}
-  </div>
-</div>
