+++ app/components/frost-object-browser-inline.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-frost-object-browser/components/frost-object-browser-inline'
