--- app/components/frost-object-browser-slots.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-frost-object-browser/components/frost-object-browser-slots'
