+++ tests/dummy/app/router.js
@@ -7,7 +7,7 @@ var Router = Ember.Router.extend({
 
 Router.map(function () {
   this.route('demo', { path: '/' })
-  this.route('slots')
+  this.route('inline')
   this.route('details')
 })
