+++ tests/dummy/app/styles/app.scss
@@ -40,7 +40,7 @@
   }
 }
 
-.dummy-slots-content-view {
+.dummy-content-view {
   flex: 1;
   display: flex;
   flex-direction: column;
